+++ src/components/layout/header/Header.tsx
@@ -1,9 +1,21 @@
 /**
  * Application Header
  */
+import React from 'react';
+import Link from 'next/link';
+// Application
+import { ToggleTheme } from 'components/theme/ToggleTheme';
+import TitleAndMetaTags from './TitleAndMetaTags';
 
-const Header = () => {
-  // TODO:
+const Header = (): JSX.Element => {
+  return (
+    <>
+      <TitleAndMetaTags />
+      <ul>
+        
+      </ul>
+    </>
+  );
 }
 
 export default Header;
